Feature: manage the deductibles

+++ app/Http/Controllers/DeductibleController.php
@@ -1,8 +1,9 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Http\Request;
-use App\Models\Company;
+use App\Models\Detail;
 use App\Models\Deductible;
 
 use Validator;
@@ -19,50 +20,52 @@ class DeductibleController extends Controller
         //
     }
 
-    public function create($id_company, Request $request) {
-    
+    public function create($id_detail, Request $request)
+    {
+
         $validator = Validator::make($request->all(), [
-            'is_saving' => 'required',
             'option' => 'required',
             'applicable' => 'required'
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return response(['message' => 'Validation errors', 'errors' =>  $validator->errors(), 'status' => false], 422);
         }
 
         $input = $request->all();
 
         $deductible = new Deductible($input);
-
-        $company = Company::find($id_company);
-        $company->deductibles()->save($deductible);
-
-        return response(['status'=>true, 'message'=>'save', 'data'=>[]]);
-
+        $detail = Detail::find($id_detail);
+        if ($detail) {
+            $detail->deductibles()->save($deductible);
+            return response(['status' => true, 'message' => 'save', 'data' => $deductible]);
+        }
+        return response(['status' => false, 'message' => 'Cannot save the register', 'data' => []]);
     }
 
-    public function delete($id_company, $id_deductible) {
+    public function delete($id_deductible)
+    {
 
-        $company = Company::find($id_company);
-        $deductible = $company->deductibles()->find($id_deductible);
+        $deductible = Deductible::find($id_deductible);
 
-        if(!$deductible->delete()) {
+        if ($deductible) {
+            if (!$deductible->delete()) {
+                return response([
+                    'status' => false,
+                    'message' => 'Fail to delete',
+                    'data' => []
+                ]);
+            }
             return response([
-                'status' => false,
-                'message' => 'Fail to delete',
+                'status' => true,
+                'message' => 'Oh yeah!',
                 'data' => []
             ]);
         }
-
         return response([
-            'status' => true,
-            'message' => 'Oh yeah!',
+            'status' => false,
+            'message' => 'Try again.',
             'data' => []
         ]);
-
     }
-
-
-
 }
